raspberry-script/requirements.txt
aranet4
pandas
google-cloud-bigquery
pyarrow
python-dotenv